+++ packages/component-library/stories/LineChart.story.js
@@ -12,7 +12,7 @@ import {
 import { LineChart, SimpleLegend, civicFormat } from "../src";
 import { getKeyNames } from "./shared";
 import notes from "./lineChart.notes.md";
-import { CrazyVisualizationTheme } from "../src/_Themes/index";
+import { VictoryCrazyTheme } from "../src/_Themes/index";
 
 const GROUP_IDS = {
   LABELS: "Labels",
@@ -150,7 +150,7 @@ export default () =>
             yLabel={yLabel}
             xNumberFormatter={x => civicFormat[optionSelectX](x)}
             yNumberFormatter={y => civicFormat[optionSelectY](y)}
-            theme={CrazyVisualizationTheme}
+            theme={VictoryCrazyTheme}
           />
         );
       },
